# flask_open_api/swagger_server/models/order.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.position_in_order import PositionInOrder  # noqa: F401,E501
from swagger_server import util


class Order(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, positions: List[PositionInOrder]=None, ship_date: datetime=None, price: float=None, status: str=None, complete: bool=None):  # noqa: E501
        """Order - a model defined in Swagger

        :param id: The id of this Order.  # noqa: E501
        :type id: int
        :param positions: The positions of this Order.  # noqa: E501
        :type positions: List[PositionInOrder]
        :param ship_date: The ship_date of this Order.  # noqa: E501
        :type ship_date: datetime
        :param price: The price of this Order.  # noqa: E501
        :type price: float
        :param status: The status of this Order.  # noqa: E501
        :type status: str
        :param complete: The complete of this Order.  # noqa: E501
        :type complete: bool
        """
        self.swagger_types = {
            'id': int,
            'positions': List[PositionInOrder],
            'ship_date': datetime,
            'price': float,
            'status': str,
            'complete': bool
        }

        self.attribute_map = {
            'id': 'id',
            'positions': 'Positions',
            'ship_date': 'shipDate',
            'price': 'Price',
            'status': 'status',
            'complete': 'complete'
        }
        self._id = id
        self._positions = positions
        self._ship_date = ship_date
        self._price = price
        self._status = status
        self._complete = complete

    @classmethod
    def from_dict(cls, dikt) -> 'Order':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Order of this Order.  # noqa: E501
        :rtype: Order
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Order.


        :return: The id of this Order.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Order.


        :param id: The id of this Order.
        :type id: int
        """

        self._id = id

    @property
    def positions(self) -> List[PositionInOrder]:
        """Gets the positions of this Order.


        :return: The positions of this Order.
        :rtype: List[PositionInOrder]
        """
        return self._positions

    @positions.setter
    def positions(self, positions: List[PositionInOrder]):
        """Sets the positions of this Order.


        :param positions: The positions of this Order.
        :type positions: List[PositionInOrder]
        """

        self._positions = positions

    @property
    def ship_date(self) -> datetime:
        """Gets the ship_date of this Order.


        :return: The ship_date of this Order.
        :rtype: datetime
        """
        return self._ship_date

    @ship_date.setter
    def ship_date(self, ship_date: datetime):
        """Sets the ship_date of this Order.


        :param ship_date: The ship_date of this Order.
        :type ship_date: datetime
        """

        self._ship_date = ship_date

    @property
    def price(self) -> float:
        """Gets the price of this Order.


        :return: The price of this Order.
        :rtype: float
        """
        return self._price

    @price.setter
    def price(self, price: float):
        """Sets the price of this Order.


        :param price: The price of this Order.
        :type price: float
        """

        self._price = price

    @property
    def status(self) -> str:
        """Gets the status of this Order.

        Order Status  # noqa: E501

        :return: The status of this Order.
        :rtype: str
        """
        return self._status

    @status.setter
    def status(self, status: str):
        """Sets the status of this Order.

        Order Status  # noqa: E501

        :param status: The status of this Order.
        :type status: str
        """
        allowed_values = ["placed", "approved", "delivered"]  # noqa: E501
        if status not in allowed_values:
            raise ValueError(
                "Invalid value for `status` ({0}), must be one of {1}"
                .format(status, allowed_values)
            )

        self._status = status

    @property
    def complete(self) -> bool:
        """Gets the complete of this Order.


        :return: The complete of this Order.
        :rtype: bool
        """
        return self._complete

    @complete.setter
    def complete(self, complete: bool):
        """Sets the complete of this Order.


        :param complete: The complete of this Order.
        :type complete: bool
        """

        self._complete = complete