# flask_open_api/swagger_server/models/pizza.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Pizza(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, name: str=None, photo_urls: List[str]=None,price: float=None):  # noqa: E501
        """Pizza - a model defined in Swagger

        :param id: The id of this Pizza.  # noqa: E501
        :type id: int
        :param name: The name of this Pizza.  # noqa: E501
        :type name: str
        :param photo_urls: The photo_urls of this Pizza.  # noqa: E501
        :type photo_urls: List[str]
        """
        self.swagger_types = {
            'id': int,
            'name': str,
            'photo_urls': List[str],
            'price':float
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'photo_urls': 'photoUrls',
            'price':'price'
        }
        self._id = id
        self._name = name
        self._photo_urls = photo_urls
        self._price=price

    @classmethod
    def from_dict(cls, dikt) -> 'Pizza':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Pizza of this Pizza.  # noqa: E501
        :rtype: Pizza
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Pizza.


        :return: The id of this Pizza.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Pizza.


        :param id: The id of this Pizza.
        :type id: int
        """

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Pizza.


        :return: The name of this Pizza.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Pizza.


        :param name: The name of this Pizza.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name
    @property
    def price(self) -> float:

        return self._price

    @price.setter
    def price(self, price: float):

        if price is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._price = price



    @property
    def photo_urls(self) -> List[str]:
        """Gets the photo_urls of this Pizza.


        :return: The photo_urls of this Pizza.
        :rtype: List[str]
        """
        return self._photo_urls

    @photo_urls.setter
    def photo_urls(self, photo_urls: List[str]):
        """Sets the photo_urls of this Pizza.


        :param photo_urls: The photo_urls of this Pizza.
        :type photo_urls: List[str]
        """
        if photo_urls is None:
            raise ValueError("Invalid value for `photo_urls`, must not be `None`")  # noqa: E501

        self._photo_urls = photo_urls