# flask_open_api/swagger_server/models/pizzas.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.pizza import Pizza  # noqa: F401,E501
from swagger_server import util


class Pizzas(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, all_pizzas: List[Pizza]=None):  # noqa: E501
        """Pizzas - a model defined in Swagger

        :param all_pizzas: The all_pizzas of this Pizzas.  # noqa: E501
        :type all_pizzas: List[Pizza]
        """
        self.swagger_types = {
            'all_pizzas': List[Pizza]
        }

        self.attribute_map = {
            'all_pizzas': 'AllPizzas'
        }
        self._all_pizzas = all_pizzas

    @classmethod
    def from_dict(cls, dikt) -> 'Pizzas':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Pizzas of this Pizzas.  # noqa: E501
        :rtype: Pizzas
        """
        return util.deserialize_model(dikt, cls)

    @property
    def all_pizzas(self) -> List[Pizza]:
        """Gets the all_pizzas of this Pizzas.


        :return: The all_pizzas of this Pizzas.
        :rtype: List[Pizza]
        """
        return self._all_pizzas

    @all_pizzas.setter
    def all_pizzas(self, all_pizzas: List[Pizza]):
        """Sets the all_pizzas of this Pizzas.


        :param all_pizzas: The all_pizzas of this Pizzas.
        :type all_pizzas: List[Pizza]
        """

        self._all_pizzas = all_pizzas